[FE] feat: Nav 한 파일로 정리
